Overhaul of CSR functionality
