+++ include/RevFeature.h
@@ -35,15 +35,16 @@ enum RevFeatureType : uint32_t {
   RV_V        = 1 << 9,   ///< RevFeatureType: V-extension
   RV_H        = 1 << 10,  ///< RevFeatureType: H-extension
   RV_ZICBOM   = 1 << 11,  ///< RevFeatureType: Zicbom-extension
-  RV_ZICSR    = 1 << 12,  ///< RevFEatureType: Zicsr-extension
-  RV_ZIFENCEI = 1 << 13,  ///< RevFeatureType: Zifencei-extension
-  RV_ZMMUL    = 1 << 14,  ///< RevFeatureType: Zmmul-extension
-  RV_ZAAMO    = 1 << 15,  ///< RevFeatureType: Zaamo-extension
-  RV_ZALRSC   = 1 << 16,  ///< RevFeatureType: Zalrsc-extension
-  RV_ZFA      = 1 << 17,  ///< RevFeatureType: Zfa-extension
-  RV_ZFH      = 1 << 18,  ///< RevFeatureType: H-extension
-  RV_ZFHMIN   = 1 << 19,  ///< RevFeatureRtpe: Zfhmin extension
-  RV_ZTSO     = 1 << 20,  ///< RevFeatureType: Ztso-extension
+  RV_ZICNTR   = 1 << 12,  ///< RevFeatureType: Zicntr-extension
+  RV_ZICSR    = 1 << 13,  ///< RevFEatureType: Zicsr-extension
+  RV_ZIFENCEI = 1 << 14,  ///< RevFeatureType: Zifencei-extension
+  RV_ZMMUL    = 1 << 15,  ///< RevFeatureType: Zmmul-extension
+  RV_ZAAMO    = 1 << 16,  ///< RevFeatureType: Zaamo-extension
+  RV_ZALRSC   = 1 << 17,  ///< RevFeatureType: Zalrsc-extension
+  RV_ZFA      = 1 << 18,  ///< RevFeatureType: Zfa-extension
+  RV_ZFH      = 1 << 19,  ///< RevFeatureType: H-extension
+  RV_ZFHMIN   = 1 << 20,  ///< RevFeatureRtpe: Zfhmin extension
+  RV_ZTSO     = 1 << 21,  ///< RevFeatureType: Ztso-extension
 };
 
 struct RevFeature {
