+++ include/insns/Zalrsc.h
@@ -30,7 +30,7 @@ class Zalrsc : public RevExt {
 
     // Flags for LR memory load
     RevFlag flags = RevFlag::F_NONE;
-    if( sizeof( TYPE ) < sizeof( int64_t ) && R->IsRV64 )
+    if( sizeof( TYPE ) < sizeof( int64_t ) && F->IsRV64() )
       RevFlagSet( flags, RevFlag::F_SEXT64 );
     if( Inst.aq )
       RevFlagSet( flags, RevFlag::F_AQ );
@@ -39,7 +39,7 @@ class Zalrsc : public RevExt {
 
     // Where the data will eventually end up
     void* target;
-    if( sizeof( TYPE ) >= sizeof( int64_t ) || R->IsRV64 ) {
+    if( sizeof( TYPE ) >= sizeof( int64_t ) || F->IsRV64() ) {
       target = &R->RV64[Inst.rd];
     } else {
       target = &R->RV32[Inst.rd];
