+++ src/RevCore.cc
@@ -1490,7 +1490,7 @@ void RevCore::HandleRegFault( unsigned width ) {
 
   if( !feature->HasF() || RevRand( 0, 1 ) ) {
     // X registers
-    if( regFile->IsRV64 ) {
+    if( regFile->IsRV64() ) {
       regFile->RV64[RegIdx] |= RevRand( 0, ~( ~uint64_t{ 0 } << width ) );
     } else {
       regFile->RV32[RegIdx] |= RevRand( 0, ~( ~uint32_t{ 0 } << width ) );
@@ -1801,7 +1801,7 @@ bool RevCore::ClockTick( SST::Cycle_t currentCycle ) {
       );
 #endif
       ++Stats.retired;
-      ++RegFile->InstRet;
+      RegFile->IncrementInstRet();
 
       // Only clear the dependency if there is no outstanding load
       if( ( RegFile->GetLSQueue()->count( LSQHash( Pipeline.front().second.rd, InstTable[Pipeline.front().second.entry].rdClass, HartID ) ) ) == 0 ) {
